requirements.txt
streamlit
ultralytics
tensorflow
opencv-python
matplotlib
Pillow
gdown
